+++ src/types/auth.ts
@@ -1,3 +1,5 @@
+import type { User } from "./user";
+
 export interface LoginFormValues {
   email: string;
   password: string;
@@ -9,4 +11,22 @@ export interface SignUpFormValues {
   passwordConfirm: string;
   name: string;
   introduce: string;
+}
+
+export interface LoginResult {
+  ok: boolean;
+  user?: User;
+}
+
+export interface SignUpResult {
+  ok: boolean;
+}
+
+export interface OAuthTokenPair {
+  accessToken: string;
+  refreshToken: string;
+}
+
+export interface OAuthLoginResult extends OAuthTokenPair {
+  user: User;
 }
